requirements.txt
flask
flask_session
gunicorn
sqlalchemy
google-cloud-speech
boto3
streamlit
